Bow And Arrow System - Shoot and Damage

+++ Assets/RangedAmmoItem.cs
@@ -19,7 +19,7 @@ public class RangedAmmoItem : Item
     public int currentAmount = 99;
 
     [Header("Ammo Base Damage")]
-    public float physicalDamage = 50;
+    public int physicalDamage = 50;
     //Magic Damage
     //Fire Damage
     //Dark Damage
